Read_Me
My name is Anthony Perez, I am 34 years old married with one boy who is two years old and another that is a girl named Sophia and she will be here any day now. I currently work with M.D.A on the I.T. move team but I am looking to better myself so I can be a valuable asset to a great company.